Remove default value for editor value. (#857)
Currently react-ace sets the default value for the editor value to "".
This means that, anytime the editor is rerendered, any changes from the
user are lost if the parent doesn't pass a value component in. This
prevents the component from being used in an uncontrolled context.

This patch removes the default "" value for the editor value, allowing
react-ace to function properly as a uncontrolled component. All tests
are passing, but the documentation should probably be updated to reflect
this change.